moving isHotfixEnabled to the JiraComponent
